+++ tests/test_compute.py
@@ -109,7 +109,7 @@ def test_basic_equations(eq, dtype, strip_exponent):
     x = np.einsum(eq, *arrays)
     y = ctg.einsum(eq, *arrays, strip_exponent=strip_exponent)
     if strip_exponent:
-        y = y[0] * 10**y[1]
+        y = y[0] * 10 ** y[1]
     rtol = 5e-3 if dtype in ("float32", "complex64") else 5e-6
     assert_allclose(x, y, rtol=rtol)